requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
rasterio>=1.3.0
albumentations>=1.3.0
ultralytics>=8.0.0
tqdm>=4.65.0
pyyaml>=6.0
wandb>=0.15.0
grad-cam>=1.4.0
scipy>=1.11.0
pandas>=2.0.0
streamlit>=1.33.0
